backend/requirements.txt
# FastAPI Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
python-multipart>=0.0.6
aioredis>=2.0.0
celery>=5.3.0
redis>=5.0.0

# Shared with Flask
werkzeug>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
